pyproject.toml
[tool.poetry]
name = "bfg-unpacker"
version = "0.1.0"
description = ""
authors = ["gmh4589"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pyqt6 = "^6.6.1"
icecream = "^2.1.3"
ffmpeg = "^1.4"
vpk = "^1.4.0"
pandas = "^2.2.0"
pillow = "^10.2.0"
jinja2 = "^3.1.3"
psutil = "^5.9.8"
requests = "^2.31.0"
pyinstaller = "^6.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
